ms_assurance_agent/requirements.txt
openai 
python-dotenv 
langchain 
faiss-cpu 
pandas 
tabulate 
langchain-openai
fastapi==0.115.12
uvicorn==0.34.1
azure-storage == 0.37.0